requirement.txt
flask==2.3.2
tensorflow==2.10.0
numpy==1.23.5
pandas==1.5.3
pillow==9.5.0
